Rename Cache policy `prefetch` to `fetch` since
Rename the cache policy fetch variable because since events are only
viewable within the first 30 seconds, there really is only a small
window of fetching at all. That is, you either fetch or don't, and
a pre-fetch isn't really the right naming.

This is a breaking API change, however nothing is using this API yet.